Implement topBar.borderColor and topBar.borderHeight

+++ lib/android/app/src/main/java/com/reactnativenavigation/parse/parsers/FractionParser.java
@@ -7,6 +7,6 @@ import org.json.JSONObject;
 
 public class FractionParser {
     public static Fraction parse(JSONObject json, String fraction) {
-        return json.has(fraction) ? new Fraction(json.optInt(fraction)) : new NullFraction();
+        return json.has(fraction) ? new Fraction(json.optDouble(fraction)) : new NullFraction();
     }
 }
